adding longs ids

+++ app/src/main/java/com/example/pam_project/db/entities/ListEntity.java
@@ -14,16 +14,16 @@ public class ListEntity {
     @ColumnInfo(name = "name")
     public String name;
 
-    public int categoryId;
+    public long categoryId;
 
-    public ListEntity(long id, String name, int categoryId) {
+    public ListEntity(long id, String name, long categoryId) {
         this.id = id;
         this.name = name;
         this.categoryId = categoryId;
     }
 
     @Ignore
-    public ListEntity(String name, int categoryId) {
+    public ListEntity(String name, long categoryId) {
         this.name = name;
         this.categoryId = categoryId;
     }
